+++ src/main/java/com/agency04/sbss/pizza/PizzaConfig.java
@@ -5,12 +5,10 @@ import com.agency04.sbss.pizza.service.impl.PizzeriaMaslina;
 import org.springframework.context.annotation.*;
 
 @Configuration
-public class PizzaConfig {
-
+public class PizzaConfig  {
     @Bean
     @Primary
     public PizzeriaService pizzeria(){
         return new PizzeriaMaslina();
     }
-
 }
