+++ src/main/java/com/agency04/sbss/pizza/model/Margherita.java
@@ -2,6 +2,8 @@ package com.agency04.sbss.pizza.model;
 
 public class Margherita implements Pizza {
 
+    public Margherita(){}
+
     @Override
     public String getName() {
         return "Margherita";
