+++ src/main/java/com/agency04/sbss/pizza/model/QuattroStagioni.java
@@ -4,7 +4,7 @@ public class QuattroStagioni implements Pizza {
 
     @Override
     public String getName() {
-        return "Quattro Stagioni";
+        return "QuattroStagioni";
     }
 
     @Override
