+++ src/main/resources/application.properties
@@ -1,3 +1,5 @@
 name = Maslina 
 address = Put Bokanjca
 phoneNumber = 0913211239
+
+spring.jackson.serialization.indent_output = true
