+++ src/main/resources/static/index.html
@@ -8,9 +8,9 @@
 <body>
 <h1>Pizza delivery</h1>
 <br>
-<a href="http://localhost:8080/api/pizzeria/menu">Get pizzeria menu</a><br>
-<a href="http://localhost:8080/api/pizzeria">Get pizzeria information</a><br>
-<a href="http://localhost:8080/api/customer">Get list of customers</a><br>
-<a href="http://localhost:8080/api/delivery/list">Get delivery list</a><br>
+<a href="/api/pizzeria/menu">Get pizzeria menu</a><br>
+<a href="/api/pizzeria">Get pizzeria information</a><br>
+<a href="/api/customer">Get list of customers</a><br>
+<a href="/api/delivery/list">Get delivery list</a><br>
 </body>
 </html>
